requirement.txt
tkinter # Standard library, no installation needed; included for clarity
google-generativeai
pyperclip
pyautogui
keyboard